Do not pre-warm partition pool
This is a first step on the road to fully concurrent reads and writes.
Up until this commit we were working with an assumption that there
should always exist at least one partition in the partition pool. This
influenced the append logic in many places, adding unnecessary
complexity and mixing two distinct features - appending data and
managing partition pool. It also had an annoying side-effect of
pre-warming each partition group's pool upon creation with a partition
with a lower ts bound of MIN_TIMESTAMP, i.e. 1970-x-x.

The new behavior is to create a partition only when needed.