Add string-based ScanFilterOps
Newly introduced: StartsWith, EndsWith, Contains and Matches (with
regular expression support).

For now, the safest bet is to almost always use Matches, as the regex
engine is so optimized that there's a good chance that it will
outperform std matchers in most cases.

All of those ops can be also used with non-string types, but with a cost
of up to two additional allocations.