Fix scan error on total ts pruning
When scanning with a ts pruning that would remove all partitions from
the scan, a "scan error" result was returned instead of an empty result
set. This commit fixes this behavior, with the caveat that the returned
data is an empty ScanResult, without any projected columns present.
This is consistent with the case when there are no partitions at all,
but is different from non-pruned scans that result in an empty data set,
where the result is a ScanResult with projected columns which contain no
data.