Introduce ThingWithTempDir test helper
TempDir used in test code is very tightly coupled with Catalog, but
there's no language-level connection between their lifetimes.
Dropping TempDir when Catalog is still alive will cause weird
errors/races, and this helper struct is an attempt to fix that.
As it implements Deref, it can be used as a drop-in replacement for
wrapped value.
This will allow to eventually and gradually drop part of test macros
in favor of simple functions.